Rough Assignment Submission
I still need to clean up a few things, but it works overall!
Changes in Database.java are all SQL-oriented - they are just for
debugging, and since we're working with SQL, these are a little nicer to
look at.
Added parsing functionality to JSONRT.java.
The old insertStudents() in TableFormat.js is still there - not too much
different from what's not commented out.