Student, course, and schedule requests done!
The HTML is still a little wacky because the table is formatting a
little weirdly... But that's an easy fix - it'll just take a bit of
looking around. The server-side code is all done, and the client-side
just needs a small fix to be completely done. I also added a button that
clears the output div, so that you don't have to refresh the page every
time you want something clean to look at.