Converted to Maven Project
Converted the SEM Console Code into a Maven Project. This took the manual adding of dependencies and artifacts for every library he used. From now on the project to edit is called SemConsoleWebApp and is in the SemRoomLiveFeed folder. The SEM-Console-master folder will serve as a backup copy of the original code