Basic Image Storage Functionality
SQL database is up and running. Database java class to take file objects and store them in the SQL database. Function to retrieve images based on index. Modified index with script tags to allow for javascript to take effect. Added a button for testing purposes in the index page. Created a javascript file with a simple get route for gallery testing. This is to avoid github conflicts until something operational is put together. Currently, this allows us to grab the file input stream from a SEM png photo and return it to javascript. Next steps: figure out how to turn that array into an image and figure out why this file input stream retrieval is unbelievably slow.